Repackaging. Capture incoming Request attributes in TestClient
