+++ zio-http/src/test/scala/zio/http/TestClientSpec.scala
@@ -9,7 +9,7 @@ object TestClientSpec extends ZIOSpecDefault {
   def spec = suite("TestClientSpec")(
     test("test") {
       for {
-        _ <- ZIO.serviceWithZIO[Client](_.request(Request(url = URL(Path.root, Location.Absolute(Scheme.HTTP, "localhost", port = 8080))))).ignore
+        _ <- ZIO.serviceWithZIO[Client](_.request(Request(url = URL(Path.root / "foo", Location.Absolute(Scheme.HTTP, "localhost", port = 8080))))).ignore
         middleTraffic <- TestClient.interactions().debug
         _ <- ZIO.serviceWithZIO[Client](_.request(Request(url = URL(Path.root / "users", Location.Absolute(Scheme.HTTP, "localhost", port = 8080))))).ignore
         finalTraffic <- TestClient.interactions().debug
